Move handlers from main to modules
